refactor: PartyStatus를 Party 도메인 아래로 변경
