+++ src/main/java/umc/cockple/demo/domain/bookmark/service/BookmarkCommandService.java
@@ -23,7 +23,7 @@ import umc.cockple.demo.domain.party.exception.PartyErrorCode;
 import umc.cockple.demo.domain.party.exception.PartyException;
 import umc.cockple.demo.domain.party.repository.PartyRepository;
 import umc.cockple.demo.global.enums.PartyOrderType;
-import umc.cockple.demo.global.enums.PartyStatus;
+import umc.cockple.demo.domain.party.enums.PartyStatus;
 
 @Service
 @Transactional
