+++ src/main/java/umc/cockple/demo/domain/party/domain/Party.java
@@ -11,6 +11,7 @@ import umc.cockple.demo.domain.member.domain.Member;
 import umc.cockple.demo.domain.member.domain.MemberParty;
 import umc.cockple.demo.domain.party.dto.PartyCreateDTO;
 import umc.cockple.demo.domain.party.enums.ActiveDay;
+import umc.cockple.demo.domain.party.enums.PartyStatus;
 import umc.cockple.demo.global.common.BaseEntity;
 import umc.cockple.demo.global.enums.*;
